requirements.txt
mysql-connector-python~=9.2.0
PySide6~=6.8.2.1